[[ Bug 18277 ]] Translate NSRect <-> MCRectangle correctly
Previously the height of the desktop was presumed to be the height
of the primary screen or [[NSScreen screens] objectAtIndex:0] this
is not correct because the user may have other monitors attached
that have their displays positioned above the main screen.

An example would be a user with a laptop and a plugged in screen
where the laptop is positioned below the secondary monitor both
physically and via the display setup. Currently in such a setup
stack rects and locations are not reported correctly because
the translation from bottom left origin to top right origin
is done using the wrong desktop height as the topmost monitor
always has a 0 y origin regardless of whether it is the primary
(menubar) monitor or not.